Config script is a gulp task, update README
Converted the config script to a gulp task. Updated the README to
include general installation info, extra info for the new gulp task, and
some minor spell checks.
