+++ ideas.js
@@ -1,0 +1,13 @@
+var ideas = [{
+  "key":0,
+  "_id":"idea_0",
+  "ideaId":0,
+  "title":"ROKStarter",
+  "description":"A Platform for Innovation that promotes collaboration by providing visibility to ideas.",
+  "author":"The Cleveland Innovation Challenge Team",
+  "likes":0,
+  "comments":[],
+  "backs":[]
+}];
+
+exports.ideas = ideas;
